services/ui/requirements.txt
fastapi==0.114.1
uvicorn[standard]==0.30.6
paho-mqtt==1.6.1
